Refactor: Refactoring entities to BaseEntity
-(User, Channel, Message) extends BaseEntity

-공통 속성 BaseEntity로 리팩토링

-Override toString()
