+++ src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,0 +1,31 @@
+package com.sprint.mission.discodeit.entity;
+
+import java.util.UUID;
+
+public abstract class BaseEntity {
+    protected final UUID id;
+    protected final Long createdAt;
+    protected Long updatedAt;
+
+    public BaseEntity() {
+        this.id = UUID.randomUUID();
+        this.createdAt = System.currentTimeMillis();
+        this.updatedAt = createdAt;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public Long getCreatedAt() {
+        return createdAt;
+    }
+
+    public Long getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt() {
+        this.updatedAt = System.currentTimeMillis();
+    }
+}
