+++ src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -1,41 +1,29 @@
 package com.sprint.mission.discodeit.entity;
 
-import java.util.UUID;
-
-public class Channel {
-    private final UUID id;
+public class Channel extends BaseEntity {
     private String name;
-    private final Long createdAt;
-    private Long updatedAt;
 
     public Channel(String name) {
+        super();
         this.name = name;
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = createdAt;
     }
 
     public String getName() {
         return name;
     }
 
-    public UUID getId() {
-        return id;
-    }
-
-    public Long getCreatedAt() {
-        return createdAt;
-    }
-
-    public Long getUpdatedAt() {
-        return updatedAt;
-    }
-
     public void setName(String name) {
         this.name = name;
+        setUpdatedAt();
     }
 
-    public void setUpdatedAt(Long updatedAt) {
-        this.updatedAt = updatedAt;
+    @Override
+    public String toString() {
+        return "Channel{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", createdAt=" + createdAt +
+                ", updatedAt=" + updatedAt +
+                '}';
     }
 }
