+++ src/main/java/com/sprint/mission/discodeit/entity/Message.java
@@ -1,41 +1,29 @@
 package com.sprint.mission.discodeit.entity;
 
-import java.util.UUID;
-
-public class Message {
-    private final UUID id;
+public class Message extends BaseEntity{
     private String text;
-    private final Long createdAt;
-    private Long updatedAt;
+
+    public Message(String text) {
+        super();
+        this.text = text;
+    }
 
     public String getText() {
         return text;
     }
 
-    public Message(String text) {
-        this.text = text;
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = createdAt;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public Long getCreatedAt() {
-        return createdAt;
-    }
-
-    public Long getUpdatedAt() {
-        return updatedAt;
-    }
-
     public void setText(String text) {
         this.text = text;
+        setUpdatedAt();
     }
 
-    public void setUpdatedAt(Long updatedAt) {
-        this.updatedAt = updatedAt;
+    @Override
+    public String toString() {
+        return "Message{" +
+                "id=" + id +
+                ", text='" + text + '\'' +
+                ", createdAt=" + createdAt +
+                ", updatedAt=" + updatedAt +
+                '}';
     }
 }
